+++ CHANGELOG.md
@@ -6,5 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 
 ## 2021-01-02
-Created MPA with API for stock data https://rapidapi.com/blog/react-multi-page-app/
-Added in the Chuck Norris generatorhttps://developer.okta.com/blog/2019/03/06/simple-user-authentication-in-react 
+- Created MPA with API for stock data https://rapidapi.com/blog/react-multi-page-app/
+- Added in the Chuck Norris generator (auth to come later ) https://developer.okta.com/blog/2019/03/06/simple-user-authentication-in-react 
+- Replaced Nav with Navigation component
+- Added banner to header